rclone: 初次并没有生成

+++ home/base/core/rclone.nix
@@ -38,6 +38,12 @@ in
     # 重新调整密码
     home.activation.postRclone = lib.hm.dag.entryAfter [ "sops-nix" ] ''
       RCLONE_PATH=${config.xdg.configHome}/rclone/rclone.conf
+
+      # 检查模板文件是否存在
+      if [ ! -f "${config.sops.templates."rclone.conf".path}" ]; then
+        exit 0
+      fi
+
       cp -f ${config.sops.templates."rclone.conf".path} $RCLONE_PATH
 
       RCLONE_DRIVE_DAV_PASSWORD=$(cat ${config.sops.secrets."drive/rclone/pass".path})
